Decompose Codex JSONL parser internals
